tests/requirements.txt
tinycss2==1.*
ruff==0.*
pillow==12.*